move game loading and socket listener to GamesCPanel component
