+++ frontend/src/components/Home.jsx
@@ -1,29 +1,9 @@
-import React, { useEffect } from 'react';
-import { connect } from 'react-redux';
-
-import socket from '../socket';
-import { actions } from '../store';
+import React from 'react';
 
 import GamesCPanel from './GamesCPanel';
 import UserStats from './UserStats';
 
-const Home = ({
-  loadGames,
-  updateGame,
-  addGameByOthers
-}) => {
-  useEffect(() => {
-    socket.on('game created', gameData => {
-      const { game } = gameData;
-      addGameByOthers(game);
-    });
-    socket.on('game joined', gameData => {
-      updateGame(gameData.game);
-    });
-
-    loadGames();
-  }, []);
-
+const Home = () => {
   return (
     <div id="home">
       <UserStats />
@@ -32,17 +12,4 @@ const Home = ({
   );
 };
 
-const mapDispatchToProps = dispatch => ({
-  addGameByOthers(game) {
-    dispatch(actions.addGame(game));
-  },
-  loadGames() {
-    dispatch(actions.getUser());
-    dispatch(actions.getGames());
-  },
-  updateGame(updatedGame) {
-    dispatch(actions.updateGame(updatedGame));
-  }
-});
-
-export default connect(null, mapDispatchToProps)(Home);
+export default Home;
